Handle 'Accept: application/json' messages as JSON-API reqs by default (#169)
Adapt the JsonApiRequestProcessor and the CrnkServlet to accept HTTP requests with 'Accept: application/json' (instead of 'application/vnd.api+json') too. This makes it easier to integrate with
tools that are not aware of JSON-API specifically.

To revert to the old behaviour, enable the configuration property 'crnk.config.resource.request.rejectPlainJson'.