Introduced @JsonApiResourceField.patchStrategy (#387)
PATCH is specified to not touch existing attributes that have not been uploaded/updated with the PATCH request. For basic attributes this is simple. For complex ones like nested POJOs and general JSON data structures, the situation is more complex and depends on the use case. In some scenarios it desired to just replace the entire data structures. In other use cases one may want to merge existing and new data (when uploading a partially updated nested object). @JsonApiResourceField.patchStrategy allows to do exactly that.  By default it will merge the data (as in the past).