initial presentation model for revised crnk-ui #517
ExplorerElement repository with path `ui/explorer` builds up a presentation model for
crnk-ui (and potentially any other kind of related UI). Computing the model
on the backend-side allows easy customization and avoids the complexity
of iterating the meta model on the frontend side.

UIModelConfig.setService(...) allows to pass a set of services where
the UI/presentation is supposed to connect to. By default just the local
in-memory instance, but can also work with multiple remote repositories
using crnk-client.