use @JsonApiExposed(false) for client repository stubs
by default client stubs should not be exposed on the
server-side (e.g. when setup and discovered by dependency injection).
Most likely the server wants to use the stubs to call other services,
not expose the same services again.

For this purpose the @JsonApiExposed logic needed to improve to detect
wrappers and dynamic stubs need to implement the Wrapper contract.