Allows for configuration of openapi output format and sort order, as well as improves speed and memory consumption of existing openapi comparison tests (#585)
Allows for configuration of openapi output format and sort  order, as well as improves speed and memory consumption of existing  openapi comparison tests Prior to this PR, the integration tests compared the OpenAPI objects  in memory. This was slow and consumed a ton of memory, especially in the  case of test failure. By adding the ability to sort the yaml output, we can now generate  yaml files deterministically, which allows us to compare them line by  line, which is much faster, and easier to debug in failure cases. Output formats and tests were added as well.